Updates routes and layout to make more sense. Moves chat and channels to their own nested routes.
