+++ WebApp.Client/src/routes/chat/+page.svelte
@@ -1,0 +1,1 @@
+<div class="text-2xl text-center">Placeholder</div>
